components/update_accordion_panel/requirements.txt

altair
folium
matplotlib
numpy
pandas
plotly
plotnine
requests
seaborn
shinywidgets
